requirements.txt
agno
openai
asana
requests
PyPDF2
pdfplumber
python-dotenv
yfinance
 